+++ CDJSVis/visclibs/bonds.h
@@ -1,3 +1,5 @@
 
 int calculateBonds(int, int *, double *, int *, int, double *, double *, double, int, double *, int *, double *, 
                    double *, int *, int *, double *, int *);
+
+int calculateDisplacementVectors(int, int *, double *, double *, double *, int *, double *);